correction facture, frais suppl, details resa

+++ src/Entity/Paiement.php
@@ -24,7 +24,7 @@ class Paiement
     private $reservation;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $montant;
 
@@ -82,12 +82,12 @@ class Paiement
         return $this;
     }
 
-    public function getMontant(): ?int
+    public function getMontant(): ?float
     {
         return $this->montant;
     }
 
-    public function setMontant(int $montant): self
+    public function setMontant(float $montant): self
     {
         $this->montant = $montant;
